Fixed index signature issue
