--- src/app/screens/Home/constants.ts
@@ -1,31 +1,0 @@
-import LibraryActiveIcon from '@assets/icLibraryActive.png';
-import LibraryIcon from '@assets/icLibrary.png';
-import MyRentalsActiveIcon from '@assets/icMyRentalsActive.png';
-import MyRentalsIcon from '@assets/icMyRentals.png';
-
-interface TabIcon {
-  focused: HTMLImageElement;
-  unfocused: HTMLImageElement;
-}
-
-interface TabBar {
-  Library: TabIcon;
-  Rentals: TabIcon;
-}
-
-const LIBRARY: TabIcon = {
-  focused: LibraryActiveIcon,
-  unfocused: LibraryIcon
-};
-
-const RENTALS: TabIcon = {
-  focused: MyRentalsActiveIcon,
-  unfocused: MyRentalsIcon
-};
-
-export const getIcon = (icon: TabIcon, focus: Boolean) => (focus ? icon.focused : icon.unfocused);
-
-export const ROUTES_ICONS: any = {
-  Library: LIBRARY,
-  Rentals: RENTALS
-};
