--- src/app/screens/Home/styles.ts
@@ -1,7 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-  container: {
-    paddingTop: 60
-  }
-});
