+++ src/interfaces/tab.ts
@@ -1,0 +1,8 @@
+export interface TabIcon {
+  focused: HTMLImageElement;
+  unfocused: HTMLImageElement;
+}
+
+export interface TabRoutes {
+  [key: string]: TabIcon;
+}
